programmers-intellij/src/test/java/p42578/SolutionTest.java
package p42578;

import org.junit.Test;

import static org.assertj.core.api.Assertions.assertThat;

public class SolutionTest {
    Solution problem = new Solution();

    @Test
    public void test1() {
        String[][] clothes = {{"yellow_hat", "headgear"}, {"blue_sunglasses", "eyewear"}, {"green_turban", "headgear"}};
        assertThat(problem.solution(clothes)).isEqualTo(5);
    }

    @Test
    public void test2() {
        String[][] clothes = {{"crow_mask", "face"}, {"blue_sunglasses", "face"}, {"smoky_makeup", "face"}};
        assertThat(problem.solution(clothes)).isEqualTo(3);
    }
}